chatbot/requirements.txt
langchain-google-genai
langchain-core
streamlit
python-dotenv 